backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
numpy==1.24.3
opencv-python==4.8.1.78
tensorflow==2.14.0
pandas==2.1.1
faiss-cpu==1.7.4
